Add scan_match_index & pattern_not_found exception
Now `sapeaob::scan_match` uses `sapeaob::scan_match_index` under the
hood then it sums the pointer of the array to get the final result.
`sapeaob::scan_match_index` now can throw an exception when it doesn't
finds the value in interest, which helps when the pattern is at the
beginning of the array (which is unusual but it's a corner case).

Modified the respective tests to catch the exception as well.