internal: added step_calculator
step_calculator does a bit of logical magic which basically allows me to
find the nearest byte equal to the first byte of the pattern in a 8-byte
manner. i.e. if the current lookup is 0xAABBCCDD EEFFAABB and the first
byte of the pattern is 0xEE, it will immediately skip the first 4 bytes
when scanning. If the first byte is not present in the 8-byte type,
it'll skip the 8 bytes too.

This has an improvement of speed of around 20%.

Used this as a  ref:
http://www.icodeguru.com/Embedded/Hacker%27s-Delight/043.htm