internal: merge_bytes now can take one argument.
Now merge_bytes can take one argument when the size of TargetType is bigger than
one. When the size of TargetType is one it doesn't make sense to use
this function so we check it with static_assert.

Changed the if constexpr of the other case too since we can do a
static_assert to avoid compilation in the first place.