API: Now explicitly asks for uint8_t or enum op.
We restrict the type of the pattern struct by using static assert,
checking the Pattern is either an enum op (which only contains `ANY`) or
a value between 0x00 and 0xFF. Since in C++17 floats aren't allowed as a
constexpr we should be fine.

Made all the necessary changes to adopt this modification.